Add observe notification logic.
Add more properties to the arguments for post_notification for the command type, timeout and should wait for response. If we should wait for response we should add timeout to the socket. This means that we will check if there is response before using recv. The code for the timeout is stolen from here http://stackoverflow.com/questions/30395258/setting-timeout-to-recv-function